Create common Gemini ChatModel and EmbeddingModel code
