+++ model-providers/google/vertex-ai-gemini/runtime/src/main/java/io/quarkiverse/langchain4j/vertexai/runtime/gemini/VertexAiGeminiChatLanguageModel.java
@@ -1,68 +1,30 @@
 package io.quarkiverse.langchain4j.vertexai.runtime.gemini;
 
-import static dev.langchain4j.data.message.AiMessage.aiMessage;
-import static dev.langchain4j.internal.Utils.getOrDefault;
-
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.time.Duration;
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.TimeUnit;
 
-import org.jboss.logging.Logger;
 import org.jboss.resteasy.reactive.client.api.LoggingScope;
 
-import dev.langchain4j.agent.tool.ToolExecutionRequest;
-import dev.langchain4j.agent.tool.ToolSpecification;
-import dev.langchain4j.data.message.AiMessage;
-import dev.langchain4j.data.message.ChatMessage;
-import dev.langchain4j.model.chat.ChatLanguageModel;
-import dev.langchain4j.model.chat.listener.ChatModelErrorContext;
 import dev.langchain4j.model.chat.listener.ChatModelListener;
-import dev.langchain4j.model.chat.listener.ChatModelRequest;
-import dev.langchain4j.model.chat.listener.ChatModelRequestContext;
-import dev.langchain4j.model.chat.listener.ChatModelResponse;
-import dev.langchain4j.model.chat.listener.ChatModelResponseContext;
-import dev.langchain4j.model.chat.request.ChatRequestParameters;
 import dev.langchain4j.model.chat.request.ResponseFormat;
-import dev.langchain4j.model.chat.request.ResponseFormatType;
-import dev.langchain4j.model.chat.request.json.JsonEnumSchema;
-import dev.langchain4j.model.output.FinishReason;
-import dev.langchain4j.model.output.Response;
-import dev.langchain4j.model.output.TokenUsage;
-import io.quarkiverse.langchain4j.gemini.common.ContentMapper;
-import io.quarkiverse.langchain4j.gemini.common.FinishReasonMapper;
+import io.quarkiverse.langchain4j.gemini.common.GeminiChatLanguageModel;
 import io.quarkiverse.langchain4j.gemini.common.GenerateContentRequest;
 import io.quarkiverse.langchain4j.gemini.common.GenerateContentResponse;
-import io.quarkiverse.langchain4j.gemini.common.GenerateContentResponseHandler;
-import io.quarkiverse.langchain4j.gemini.common.GenerationConfig;
-import io.quarkiverse.langchain4j.gemini.common.SchemaMapper;
 import io.quarkus.rest.client.reactive.QuarkusRestClientBuilder;
 
-public class VertexAiGeminiChatLanguageModel implements ChatLanguageModel {
-    private static final Logger log = Logger.getLogger(VertexAiGeminiChatLanguageModel.class);
-
-    private final Double temperature;
-    private final Integer maxOutputTokens;
-    private final Integer topK;
-    private final Double topP;
-    private final ResponseFormat responseFormat;
-    private final List<ChatModelListener> listeners;
+public class VertexAiGeminiChatLanguageModel extends GeminiChatLanguageModel {
 
     private final VertxAiGeminiRestApi.ApiMetadata apiMetadata;
     private final VertxAiGeminiRestApi restApi;
 
     private VertexAiGeminiChatLanguageModel(Builder builder) {
-        this.temperature = builder.temperature;
-        this.maxOutputTokens = builder.maxOutputTokens;
-        this.topK = builder.topK;
-        this.topP = builder.topP;
-        this.responseFormat = builder.responseFormat;
-        this.listeners = builder.listeners;
+        super(builder.modelId, builder.temperature, builder.maxOutputTokens, builder.topK, builder.topP, builder.responseFormat,
+                builder.listeners);
 
         this.apiMetadata = VertxAiGeminiRestApi.ApiMetadata
                 .builder()
@@ -91,164 +53,14 @@ public class VertexAiGeminiChatLanguageModel implements ChatLanguageModel {
     }
 
     @Override
-    public dev.langchain4j.model.chat.response.ChatResponse chat(dev.langchain4j.model.chat.request.ChatRequest chatRequest) {
-        ChatRequestParameters requestParameters = chatRequest.parameters();
-        ResponseFormat effectiveResponseFormat = getOrDefault(requestParameters.responseFormat(), responseFormat);
-        GenerationConfig generationConfig = GenerationConfig.builder()
-                .maxOutputTokens(getOrDefault(requestParameters.maxOutputTokens(), this.maxOutputTokens))
-                .responseMimeType(computeMimeType(effectiveResponseFormat))
-                .responseSchema(effectiveResponseFormat != null
-                        ? SchemaMapper.fromJsonSchemaToSchema(effectiveResponseFormat.jsonSchema())
-                        : null)
-                .stopSequences(requestParameters.stopSequences())
-                .temperature(getOrDefault(requestParameters.temperature(), this.temperature))
-                .topK(getOrDefault(requestParameters.topK(), this.topK))
-                .topP(getOrDefault(requestParameters.topP(), this.topP))
-                .build();
-
-        GenerateContentRequest request = ContentMapper.map(chatRequest.messages(), chatRequest.toolSpecifications(),
-                generationConfig);
-
-        ChatModelRequest modelListenerRequest = createModelListenerRequest(request, chatRequest.messages(),
-                chatRequest.toolSpecifications());
-        Map<Object, Object> attributes = new ConcurrentHashMap<>();
-        ChatModelRequestContext requestContext = new ChatModelRequestContext(modelListenerRequest, attributes);
-        listeners.forEach(listener -> {
-            try {
-                listener.onRequest(requestContext);
-            } catch (Exception e) {
-                log.warn("Exception while calling model listener", e);
-            }
-        });
-
-        try {
-            GenerateContentResponse response = restApi.generateContent(request, apiMetadata);
-
-            String text = GenerateContentResponseHandler.getText(response);
-            List<ToolExecutionRequest> toolExecutionRequests = GenerateContentResponseHandler
-                    .getToolExecutionRequests(response);
-            AiMessage aiMessage = toolExecutionRequests.isEmpty()
-                    ? aiMessage(text)
-                    : aiMessage(text, toolExecutionRequests);
-
-            final TokenUsage tokenUsage = GenerateContentResponseHandler.getTokenUsage(response.usageMetadata());
-            final FinishReason finishReason = FinishReasonMapper.map(GenerateContentResponseHandler.getFinishReason(response));
-            final Response<AiMessage> aiMessageResponse = Response.from(aiMessage, tokenUsage);
-
-            ChatModelResponse modelListenerResponse = createModelListenerResponse(
-                    null,
-                    apiMetadata.modelId,
-                    aiMessageResponse);
-            ChatModelResponseContext responseContext = new ChatModelResponseContext(
-                    modelListenerResponse,
-                    modelListenerRequest,
-                    attributes);
-            listeners.forEach(listener -> {
-                try {
-                    listener.onResponse(responseContext);
-                } catch (Exception e) {
-                    log.warn("Exception while calling model listener", e);
-                }
-            });
-
-            return dev.langchain4j.model.chat.response.ChatResponse.builder()
-                    .aiMessage(aiMessage)
-                    .tokenUsage(tokenUsage)
-                    .finishReason(finishReason)
-                    .build();
-        } catch (RuntimeException e) {
-            ChatModelErrorContext errorContext = new ChatModelErrorContext(
-                    e,
-                    modelListenerRequest,
-                    null,
-                    attributes);
-
-            listeners.forEach(listener -> {
-                try {
-                    listener.onError(errorContext);
-                } catch (Exception e2) {
-                    log.warn("Exception while calling model listener", e2);
-                }
-            });
-
-            throw e;
-        }
-    }
-
-    @Override
-    public Response<AiMessage> generate(List<ChatMessage> messages, List<ToolSpecification> toolSpecifications) {
-        var chatResponse = chat(dev.langchain4j.model.chat.request.ChatRequest.builder()
-                .messages(messages)
-                .toolSpecifications(toolSpecifications)
-                .build());
-
-        return Response.from(
-                chatResponse.aiMessage(),
-                chatResponse.tokenUsage(),
-                chatResponse.finishReason());
-
-    }
-
-    @Override
-    public Response<AiMessage> generate(List<ChatMessage> messages) {
-        return generate(messages, Collections.emptyList());
-    }
-
-    @Override
-    public Response<AiMessage> generate(List<ChatMessage> messages, ToolSpecification toolSpecification) {
-        return generate(messages,
-                toolSpecification != null ? Collections.singletonList(toolSpecification) : Collections.emptyList());
-    }
-
-    private ChatModelRequest createModelListenerRequest(GenerateContentRequest request,
-            List<ChatMessage> messages,
-            List<ToolSpecification> toolSpecifications) {
-        var builder = ChatModelRequest.builder()
-                .model(apiMetadata.modelId)
-                .messages(messages)
-                .toolSpecifications(toolSpecifications)
-                .temperature(temperature)
-                .topP(topP)
-                .maxTokens(maxOutputTokens);
-        return builder.build();
-    }
-
-    private ChatModelResponse createModelListenerResponse(String responseId,
-            String responseModel,
-            Response<AiMessage> response) {
-        if (response == null) {
-            return null;
-        }
-
-        return ChatModelResponse.builder()
-                .id(responseId)
-                .model(responseModel)
-                .tokenUsage(response.tokenUsage())
-                .finishReason(response.finishReason())
-                .aiMessage(response.content())
-                .build();
-    }
-
-    private String computeMimeType(ResponseFormat responseFormat) {
-        if (responseFormat == null || ResponseFormatType.TEXT.equals(responseFormat.type())) {
-            return "text/plain";
-        }
-
-        if (ResponseFormatType.JSON.equals(responseFormat.type()) &&
-                responseFormat.jsonSchema() != null &&
-                responseFormat.jsonSchema().rootElement() != null &&
-                responseFormat.jsonSchema().rootElement() instanceof JsonEnumSchema) {
-            return "text/x.enum";
-        }
-
-        return "application/json";
+    protected GenerateContentResponse generateContext(GenerateContentRequest request) {
+        return restApi.generateContent(request, apiMetadata);
     }
 
     public static Builder builder() {
         return new Builder();
     }
 
-    @SuppressWarnings("OptionalUsedAsFieldOrParameterType")
     public static final class Builder {
 
         private Optional<String> baseUrl = Optional.empty();
@@ -340,4 +152,5 @@ public class VertexAiGeminiChatLanguageModel implements ChatLanguageModel {
             return new VertexAiGeminiChatLanguageModel(this);
         }
     }
+
 }
